code clean up. logic moved to service. test identified to be run on single thread.
