+++ Cargo.toml
@@ -54,6 +54,7 @@ once_cell = "1.16"
 pin-utils = "0.1.0"
 pretty_env_logger = "0.4.0"
 rand = "0.8.5"
+regex = "1.7.0"
 reqwest = { version = "0.11.11", features = ["blocking", "json"] }
 rusqlite = { version = "0.28.0", features = ["bundled"] }
 serde = { version = "1.0", features = ["derive"] }
