+++ pyrsia_cli/Cargo.toml
@@ -13,7 +13,6 @@ confy = "0.5.0"
 const_format = "0.2.26"
 futures = { version = "0.3.*"}
 lazy_static = "1.4.0"
-regex = "1.7.0"
 reqwest = { version = "0.11", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
